% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_grouped_pm_cash_flow.R
\name{append_nav_has_reported}
\alias{append_nav_has_reported}
\title{Append column if NAV has reported at end date}
\usage{
append_nav_has_reported(.data, end_date)
}
\arguments{
\item{.data}{is from get_pm_nav_daily() after filter_nav_on_dates()}

\item{end_date}{is a string (format yyyy-dd-mm)}
}
\description{
Looks for NAV at end_date and appends TRUE / FALSE column
}
