% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_globals.R
\docType{data}
\name{AZASRS__welcome__message__}
\alias{AZASRS__welcome__message__}
\title{Opening Remarks}
\format{
An object of class \code{character} of length 1.
}
\usage{
AZASRS__welcome__message__
}
\description{
Prints statements as the package loads, not really a function
}
\keyword{datasets}
