% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_globals.R
\docType{data}
\name{AZASRS_DATABASE_LOCATION}
\alias{AZASRS_DATABASE_LOCATION}
\title{Database location - should reflect production database "shared" location}
\format{An object of class \code{character} of length 1.}
\usage{
AZASRS_DATABASE_LOCATION
}
\description{
Database location - should reflect production database "shared" location
}
\examples{
"P:/IMD/2018 Database Project/Database/asrs_database.db"
}
\keyword{datasets}
