% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_benchmark_fv_index_factor.R
\name{build_benchmark_fv_index_factor}
\alias{build_benchmark_fv_index_factor}
\title{Build a tibble of benchmark index future value factor}
\usage{
build_benchmark_fv_index_factor(
  con = AZASRS_DATABASE_CONNECTION(),
  start_date = "2004-09-30",
  value_date = get_value_date(con = con),
  benchmark_daily = get_benchmark_daily_index(con = con, all_bench_types = TRUE,
    return_tibble = FALSE)
)
}
\arguments{
\item{value_date}{is the last quarter end, use for custom time ranges}

\item{benchmark_daily}{is data from get_benchmark_daily_index() and can be previously loaded and filtered}

\item{...}{grouping variables (pm_fund_id, pm_fund_portfolio, etc.)}

\item{bench_type}{is string that represents one of the types (PVT, etc.)}
}
\value{
Returns a tibble with grouping variables and all of their respective metrics
}
\description{
Build a tibble of benchmark index future value factor
}
