% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_grouped_irrs.R
\name{build_grouped_irrs}
\alias{build_grouped_irrs}
\title{Build a tibble of rolling IRRs by any grouping}
\usage{
build_grouped_irrs(
  ...,
  con = AZASRS::AZASRS_DATABASE_CONNECTION(),
  start_date = "2017-12-31",
  end_date = get_value_date(con = con),
  time_delta = "quarters",
  n_qtrs = 4,
  itd = FALSE,
  cash_adjusted = FALSE
)
}
\arguments{
\item{...}{grouping variables (pm_fund_id, pm_fund_portfolio, etc.)}

\item{start_date}{is first date}

\item{end_date}{is last date}

\item{time_delta}{is type of lag (quarters or years)}

\item{n_qtrs}{is how far back the lookback window should be (4 = 1 year, 20 = 5 years if time_delta is quarters)}

\item{itd}{whether or not you would like to have an ITD variable appended}
}
\value{
Returns a tibble with grouping variables and all of their respective metrics
}
\description{
Build a tibble of rolling IRRs by any grouping
}
