% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_modified_cash_flow.R
\name{build_modified_cash_flow}
\alias{build_modified_cash_flow}
\title{BUILD CASH FLOW TIBBLE firt for asrsMethods::irr.z}
\usage{
build_modified_cash_flow(
  ...,
  con = AZASRS_DATABASE_CONNECTION(),
  start_date = "2017-12-31",
  end_date = get_value_date(),
  return_tibble = FALSE
)
}
\arguments{
\item{...}{is the grouping set of variable(s) requested}

\item{start_date}{is the beginning date you would like the IRR to be calculated from
Should be in string format: 'yyyy-mm-dd'}

\item{end_date}{is the last date you would like the IRR to be calculated to
Should be in string format: 'yyyy-mm-dd'}
}
\description{
NAV is setup as negative at first and positive at the end with all cash flows in between where contributions are negative and distributions are positive.
}
