% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_nav_cash_flow_combined.R
\name{build_nav_cash_flow_combined}
\alias{build_nav_cash_flow_combined}
\title{Get and combine cash flows and nav according to irr.z calc}
\usage{
build_nav_cash_flow_combined(
  ...,
  con = AZASRS_DATABASE_CONNECTION(),
  start_date = "2017-12-31",
  end_date = get_value_date(con = con),
  nav_daily = get_pm_nav_daily(con = con, return_tibble = FALSE),
  cash_flow_daily = get_pm_cash_flow_daily(con = con, return_tibble = FALSE),
  itd = FALSE,
  return_tibble = FALSE
)
}
\description{
Get and combine cash flows and nav according to irr.z calc
}
