% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_privm_metrics.R
\name{build_privm_metrics}
\alias{build_privm_metrics}
\title{Build a tibble of major private market metrics including: IRR, DPI, TVPI, PME}
\usage{
build_privm_metrics(..., nav_daily = get_pm_nav_daily(),
  cf_daily = get_pm_cash_flow_daily(),
  benchmark_daily = get_benchmark_daily_index(), date_start = "1900-01-01",
  date_cutoff = as.character(lubridate::today()),
  valdate = get_value_date())
}
\arguments{
\item{...}{grouping variables (pm_fund_id, pm_fund_portfolio, etc.)}

\item{nav_daily}{is data from get_pm_nav_daily() and can be previously loaded and filtered}

\item{cf_daily}{is data from get_pm_cash_flow_daily() and can be previously loaded and filtered}

\item{date_start}{is the earliest date to be used for calculations, use for custom time ranges}

\item{date_cutoff}{is the last date to be used for PCAP, use for custom time ranges}

\item{valdate}{is the last quarter end, use for custom time ranges}
}
\value{
Returns a tibble with grouping variables and all of their respective metrics
}
\description{
Build a tibble of major private market metrics including: IRR, DPI, TVPI, PME
}
\examples{
build_privm_metrics(pm_fund_portfolio, pm_fund_id)
# pm_fund_portfolio pm_fund_id    irr     dpi   tvpi appreciation     pme irr_pcap
# <chr>             <chr>       <dbl>   <dbl>  <dbl>        <dbl>   <dbl>    <dbl>
# 1 Credit            Fund1     0.0919   0.372   1.15   437251492.   1.05        0.11
# 2 Credit            Fund2     0.159    0       1.11   562211000    1.10        NA
# 3 Credit            Fund3     0.0247   0       1.10   262390142    0.952       NA
# 4 Credit            Fund4     0.0673   0.135   1.17   173259780    1.06        NA
}
