% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_privm_p2p_irr.R
\name{build_privm_p2p_irr}
\alias{build_privm_p2p_irr}
\title{BUILD PRIVATE MARKET P2P IRR TIBBLE}
\usage{
build_privm_p2p_irr(start_date = "2017-12-31",
  end_date = get_value_date())
}
\arguments{
\item{start_date}{is the beginning date you would like the IRR to be calculated from
Should be in string format: 'yyyy-mm-dd'}

\item{end_date}{is the last date you would like the IRR to be calculated to
Should be in string format: 'yyyy-mm-dd'}
}
\description{
Creates a tibble of all Private Market IRRs found within the pm_fund_info table
}
