% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{calc_add_qtrs}
\alias{calc_add_qtrs}
\title{Calculate a future quarter date}
\usage{
calc_add_qtrs(start_date, n)
}
\arguments{
\item{start_date}{is a string (format yyyy-dd-mm)}

\item{n}{is the number of quarters to add}
}
\description{
Calculate any quarter in the future (via # of quarters)
}
