% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_finance_basics.R
\name{calc_appreciation}
\alias{calc_appreciation}
\title{Appreciation}
\usage{
calc_appreciation(cash_flow, nav)
}
\arguments{
\item{cash_flow}{is a c() of cash flows}

\item{nav}{is NAV at the beginning and end with zeros in between. Both NAV are positive.}
}
\description{
Calculates Appreciation that is NOT based off of asrsMethods library. asrsMethods needs to be updated because it does not reflect TVPI as Kerry calculates.
}
