% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{calc_fytd_metadata}
\alias{calc_fytd_metadata}
\title{Find most recent FYTD compared to end_date}
\usage{
calc_fytd_metadata(end_date)
}
\arguments{
\item{end_date}{'yyyy-mm-dd' where start_date and n_qtrs will be calculated}
}
\description{
Function returns fiscal year to date start date and number of quarters
}
