% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_grouped_irrs.R
\name{calc_grouped_irrs}
\alias{calc_grouped_irrs}
\title{Calculates IRR's given a date range}
\usage{
calc_grouped_irrs(
  ...,
  con = AZASRS_DATABASE_CONNECTION(),
  start_date = "2017-12-31",
  end_date = get_value_date(con = con),
  itd = FALSE,
  cash_adjusted = FALSE
)
}
\arguments{
\item{...}{is the grouping set of variable(s) requested}

\item{start_date}{is the beginning date you would like the IRR to be calculated from
Should be in string format: 'yyyy-mm-dd'}

\item{end_date}{is the last date you would like the IRR to be calculated to
Should be in string format: 'yyyy-mm-dd'}
}
\description{
Calculates cash flow list properly and adds NAV where required
}
\examples{
con = calc_grouped_irrs(pm_fund_portfolio, pm_fund_category, start_date = '2018-12-31')
}
