% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_grouped_pm_metrics.R
\name{calc_grouped_pm_metrics}
\alias{calc_grouped_pm_metrics}
\title{Calculate the rollup metrics}
\usage{
calc_grouped_pm_metrics(.data, ...)
}
\arguments{
\item{.data}{is from build_grouped_pm_cash_flow()}

\item{...}{aggregation choices from from pm_fund_info (i.e. pm_fund_portfolio, pm_fund_category, pm_fund_id)}
}
\description{
Calculates the major financial metrics
}
