% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finance_calculations.R
\name{calc_irr}
\alias{calc_irr}
\title{IRR}
\usage{
calc_irr(cash_flow, dates)
}
\arguments{
\item{cash_flow}{is a c() of cash_flows with the first being the earliest NAV (negative) and the last being NAV (positive)}

\item{dates}{is a c() of dates that corresponds with the cash_flow}
}
\description{
Calculates IRR based off of asrsMethods library. Can be used with tidyverse style.
}
