% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finance_calculations.R
\name{calc_irr_from_df}
\alias{calc_irr_from_df}
\title{IRR abstraction for calc_irr via a dataframe}
\usage{
calc_irr_from_df(cash_flow_df)
}
\arguments{
\item{cash_flow_df}{is a tibble with 2 columns "cash_flow" and "effective_date"}
}
\description{
Abstraction to allow us to use calc_irr from a dataframe rather than a
}
