% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_finance_basics.R
\name{calc_pme}
\alias{calc_pme}
\title{PME}
\usage{
calc_pme(distributions, contributions, nav, fv_index_factors)
}
\arguments{
\item{distributions}{are negative cash flow}

\item{contributions}{are positive cash flow}

\item{nav}{is NAV at the beginning and end with zeros in between. Both NAV are positive.}

\item{fv_index_factors}{are a c() of future value data points based off of index value divided by first index value (i.e. 1.18, 1.11, 1.05, 1.00)}
}
\description{
Calculates PME that is NOT based off of asrsMethods library. asrsMethods needs to be updated because it does not reflect TVPI as Kerry calculates.
}
