% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{calc_time_delta}
\alias{calc_time_delta}
\title{Calculate time difference}
\usage{
calc_time_delta(.data, start_date, end_date)
}
\arguments{
\item{start_date}{is the first date}

\item{end_date}{is the end date}
}
\description{
Quarters, Years, etc. need to be counted
}
