% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_finance_basics.R
\name{calc_tvpi}
\alias{calc_tvpi}
\title{TVPI}
\usage{
calc_tvpi(distributions, contributions, nav)
}
\arguments{
\item{distributions}{are positive cash flow}

\item{contributions}{are negative cash flow}

\item{nav}{is NAV at the beginning and end with zeros in between. Both NAV are positive.}
}
\description{
Calculates TVPI that is NOT based off of asrsMethods library. asrsMethods needs to be updated because it does not reflect TVPI as Kerry calculates.
}
