% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_grouped_irrs.R
\name{calculate_grouped_irr}
\alias{calculate_grouped_irr}
\title{Calculate the rollup IRR}
\usage{
calculate_grouped_irr(.data, ...)
}
\arguments{
\item{.data}{is from clean_nav_cf()}

\item{...}{aggregation choices from from pm_fund_info (i.e. pm_fund_portfolio, pm_fund_category, pm_fund_id)}
}
\description{
Calculates the IRR from a list of "cash flow" from clean_nav_cf()
}
