% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_grouped_irrs.R
\name{clean_nav_cf}
\alias{clean_nav_cf}
\title{Cleans NAV / CF filtered by date data to sum up by pm_fund_id}
\usage{
clean_nav_cf(.data, pm_fund_info)
}
\arguments{
\item{.data}{is from filter_dates()}

\item{pm_fund_info}{from get_pm_fund_info()}
}
\description{
NAV / CF joined from filter_dates()
}
