% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_grouped_irrs.R
\name{convert_start_date_nav_to_negative}
\alias{convert_start_date_nav_to_negative}
\title{Convert first NAV to negative}
\usage{
convert_start_date_nav_to_negative(.data, start_date)
}
\arguments{
\item{.data}{is from get_pm_nav_daily() after append_nav_has_reported()}

\item{start_date}{is a string (format yyyy-dd-mm)}

\item{itd}{TRUE / FALSE for itd (overwrites start_date if TRUE)}
}
\description{
Simulates cash flow at start_date from NAV
}
