% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_queries.R
\name{db_connect}
\alias{db_connect}
\title{Connect to the database}
\usage{
db_connect()
}
\arguments{
\item{db_driver}{is the driver for the type of database being accessed}

\item{db_location}{is the name of the database}
}
\value{
A connection to the database
}
\description{
This function connects to a SQLite database (this should be changed later when production is ready) but the functionality will still apply
}
