% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_queries.R
\name{db_disconnect}
\alias{db_disconnect}
\title{Disconnect from the database}
\usage{
db_disconnect(con)
}
\arguments{
\item{con}{is the current connection (likely db_connect)}
}
\value{
Disconnect from the database
}
\description{
This function connects to a SQLite database (this should be changed later when production is ready) but the functionality will still apply
}
