% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_globals.R
\name{DISCONNECT_ALL_DATABASE_CONNECTIONS}
\alias{DISCONNECT_ALL_DATABASE_CONNECTIONS}
\title{Disconnect all database connections}
\usage{
DISCONNECT_ALL_DATABASE_CONNECTIONS()
}
\description{
This will disconnect ALL database connections (use carefully, will affect all those connected)
It will send an "error" back with a 'server closed the connection unexpectedly' response. This is what you want.
}
