% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_grouped_pm_cash_flow.R
\name{filter_cf_between_dates}
\alias{filter_cf_between_dates}
\title{Collect cash flow between dates}
\usage{
filter_cf_between_dates(.data, start_date, end_date, itd)
}
\arguments{
\item{.data}{is from get_pm_cash_flow_daily()}

\item{start_date}{is a string (format yyyy-dd-mm)}

\item{end_date}{is a string (format yyyy-dd-mm)}

\item{itd}{TRUE / FALSE for itd (overwrites start_date if TRUE)}
}
\description{
Includes start and end dates, allows for ITD
}
