% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_grouped_irrs.R
\name{filter_dates}
\alias{filter_dates}
\title{Filter dates from NAV / CF joined}
\usage{
filter_dates(.data, start_date, end_date, itd, ...)
}
\arguments{
\item{.data}{is from merge_nav_and_cf()}

\item{start_date}{is a string (format yyyy-dd-mm)}

\item{end_date}{is a string (format yyyy-dd-mm)}

\item{itd}{is TRUE / FALSE for inception to date}

\item{...}{aggregation choices from from pm_fund_info (i.e. pm_fund_portfolio, pm_fund_category, pm_fund_id)}
}
\description{
NAV / CF joined from merge_nav_and_cf() needs to filter out dates IF not aggregated and not ITD. This avoids filtering out important data from roll up calculations
}
