% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_grouped_pm_cash_flow.R
\name{filter_nav_on_dates}
\alias{filter_nav_on_dates}
\title{Filter Private market NAV on dates}
\usage{
filter_nav_on_dates(.data, start_date, end_date, itd)
}
\arguments{
\item{.data}{is from get_pm_nav_daily()}

\item{start_date}{is a string (format yyyy-dd-mm)}

\item{end_date}{is a string (format yyyy-dd-mm)}

\item{itd}{TRUE / FALSE for itd (overwrites start_date if TRUE)}
}
\description{
Finds start / end date NAV and can utilize ITD
}
