% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_account_futures.R
\name{get_account_futures}
\alias{get_account_futures}
\title{Get all account_futures}
\usage{
get_account_futures(con = AZASRS_DATABASE_CONNECTION(), return_tibble = TRUE)
}
\arguments{
\item{con}{is a database connection object from AZASRS::AZASRS_DATABASE_CONNECTION()}

\item{return_tibble}{is a boolean that determines whether or not a tibble is returned instead}
}
\value{
Returns a tibble or SQL result with all pm_fund_info metadata.
}
\description{
Getting all account futures data
}
