% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_benchmark_daily_index.R
\name{get_benchmark_daily_index}
\alias{get_benchmark_daily_index}
\title{Get all benchmark_daily_index}
\usage{
get_benchmark_daily_index(
  con = AZASRS_DATABASE_CONNECTION(),
  return_tibble = FALSE
)
}
\arguments{
\item{...}{filtering expressions. i.e. effective_date >= '2018-01-01'}
}
\value{
Returns a tibble with all benchmark index data but can be filtered.
}
\description{
Gets benchmark index data by day and only allows one bench_type (i.e. SAA). Joins data from pm_fund_info, benchmark_info and filters by benchmark type. Caution: returns A LOT of data, filter for better performance.
}
\examples{
get_benchmark_daily_index(effective_date >= '2018-01-01')
# A tibble: 134,953 x 28
# benchmark_info_id effective_date index_value pm_fund_info_id benchmark_type_info_id benchmark_type pm_fund_id
# <int>                <date>               <dbl>           <int>                  <int> <chr>          <chr>
#  64                 2018-01-01           3666.              38                      1 SAA            Fund1
#  64                 2018-01-01           3666.              39                      1 SAA            Fund2
#  64                 2018-01-01           3666.              40                      1 SAA            Fund3
# ... with 134,943 more rows, and 21 more variables: pm_fund_description <chr>, pm_fund_portfolio_id <int>,
#   pm_fund_category_id <int>, pm_fund_sponsor_id <int>, pm_fund_city_id <int>, pm_fund_sector_id <int>,
#   vintage <int>, commit <dbl>, unfunded <dbl>, legacy <chr>, specialist <chr>, invest_end <chr>, term_end <chr>,
#   extension <dbl>, ext_time <dbl>, ext_used <dbl>, fee_cat <chr>, consultant <chr>, adv_board <int>, obsvr <int>,
#   fund_size_m <dbl>
}
