% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_benchmark_fund_relationship.R
\name{get_benchmark_fund_relationship}
\alias{get_benchmark_fund_relationship}
\title{Get relationships between funds and benchmarks get_benchmark_fund_relationship}
\usage{
get_benchmark_fund_relationship(
  con = AZASRS_DATABASE_CONNECTION(),

  benchmark_type = "PVT",
  all_benchmark_types = FALSE,

  return_tibble = FALSE
)
}
\arguments{
\item{con}{is the db connection, default of AZASRS_DATABASE_CONNECTION()}
}
\value{
Returns a table of relationships of ALL types, not filtered
}
\description{
Finds all data from benchmark_info_id matched with pm_fund_info_id
}
