% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_benchmark_index_daily.R
\name{get_benchmark_index_daily}
\alias{get_benchmark_index_daily}
\title{Query benchmark_index table from database}
\usage{
get_benchmark_index_daily(..., con = AZASRS_DATABASE_CONNECTION())
}
\arguments{
\item{...}{are for filtering the tables, can take values: category, portfolio, asset_class, sub_portfolio, previous_saa, sponsor, saa_benchmark, imp_benchmark, ticker, city}
}
\value{
Returns a tibble of all of the data
}
\description{
Query benchmark_index table from database
}
\examples{
get_benchmark_index_daily(effective_date >= '2018-01-01')
}
