% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_book_of_record.R
\name{get_book_of_record}
\alias{get_book_of_record}
\title{Query book_of_record_..._ table from database}
\usage{
get_book_of_record(record_type, frequency, ...,
  con = AZASRS_DATABASE_CONNECTION())
}
\arguments{
\item{record_type}{is either 'composite' or 'account'}

\item{frequency}{is either 'daily' or 'monthly'}

\item{...}{are for filtering the tables, can take values: category, portfolio, asset_class, sub_portfolio, previous_saa, sponsor, saa_benchmark, imp_benchmark, ticker, city}
}
\value{
Returns a tibble of all of the data
}
\description{
Query book_of_record_..._ table from database
}
\examples{
get_book_of_record_daily('account', 'daily', asset_class == 'Equities')
}
