% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sandbox.R
\name{get_cashflow}
\alias{get_cashflow}
\title{Query cashflow data}
\usage{
get_cashflow(from_date = "2017-01-01", short_name = "",
  return_type = "tibble")
}
\arguments{
\item{db_con}{database connection}
}
\value{
cashflow dataframe (date, shortname, amount)
}
\description{
SELECT statement to retrieve all cashflow data
}
