% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_variables.R
\name{get_filtered_benchmark}
\alias{get_filtered_benchmark}
\title{Query filtered benchmark data}
\usage{
get_filtered_benchmark(shortname, convert_365 = FALSE, return_zoo = FALSE)
}
\arguments{
\item{convert_365}{boolean to use na.approx to interpolate values with days365 function}

\item{return_zoo}{boolean to return a zoo object instead of a tibble}

\item{db_con}{database connection}
}
\value{
dataframe of shortname, longname, date, price (adds in log price of ODCE as Fixed8)
}
\description{
SELECT statement to retrieve benchmark data filtered by shortname
If 'Fixed 8' is desired, it can be entered as a shortname but it does not come from the database
Fixed 8 is calculated off of the length of ODCE returned from the database
}
