% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_variables.R
\name{get_filtered_benchmark_returns}
\alias{get_filtered_benchmark_returns}
\title{Get benchmark returns}
\usage{
get_filtered_benchmark_returns(shortnames)
}
\arguments{
\item{shortnames}{a list of shortnames c("ODCE", "Fixed 8", "RGBDAL_p250")}

\item{db_con}{database connection}
}
\value{
dataframe of benchmark returns
}
\description{
Queries benchmark with returns, add basis points by adding _p250 or _p350 to the end of each shortname
}
