% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_variables.R
\name{get_filtered_cashflow}
\alias{get_filtered_cashflow}
\title{Get cashflow}
\usage{
get_filtered_cashflow(shortname, convert_365 = FALSE, return_zoo = FALSE)
}
\arguments{
\item{shortname}{shortname}

\item{db_con}{database connection}
}
\value{
date, shortname, amount
}
\description{
Queries cashflow with filters
}
