% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_variables.R
\name{get_filtered_nav}
\alias{get_filtered_nav}
\title{Get nav values}
\usage{
get_filtered_nav(shortname)
}
\arguments{
\item{shortname}{character - shortname is the symbol}

\item{db_con}{database connection}
}
\value{
date, shortname, amount
}
\description{
Queries values with filters
}
