% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_queries.R
\name{get_info_tbl}
\alias{get_info_tbl}
\title{Query major info tables from database}
\usage{
get_info_tbl(tbl_name, ..., con = dbCon)
}
\arguments{
\item{tbl_name}{is the name of the major info table (choose one): pm_fund_info, ssbt_composite_info, account_info}

\item{...}{are for filtering the tables, can take values: category, portfolio, asset_class, sub_portfolio}

\item{con}{is the database connection (defaults to P: drive)}
}
\description{
Getting data from pm_fund_info, ssbt_composite_info, account_info with filtering possibilities
}
\examples{
get_info_tbl('pm_fund_info', sub_portfolio='Equity')
}
