% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_globals.R
\name{get_next_quarter}
\alias{get_next_quarter}
\title{Get next quarter date}
\usage{
get_next_quarter(con = AZASRS_DATABASE_CONNECTION())
}
\description{
Finds the next quarter date based off of constants table from most recent database population
}
