% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pm_fund_info.R
\name{get_pm_fund_info}
\alias{get_pm_fund_info}
\title{Query pm_fund_info table from database}
\usage{
get_pm_fund_info(...)
}
\arguments{
\item{...}{are for filtering the tables, can take values: category, portfolio, asset_class, sub_portfolio, previous_saa, sponsor, saa_benchmark, imp_benchmark, ticker, city}
}
\value{
Returns a tibble of all of the data
}
\description{
Query pm_fund_info table from database
}
\examples{
get_pm_fund_info(asset_class == 'Equities', category == 'Large')
}
