% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_csv_from_azure.R
\name{get_raw_csv_from_azure}
\alias{get_raw_csv_from_azure}
\title{Get CSV Data from Azure}
\usage{
get_raw_csv_from_azure(data_directory = "./reconciliation_data/")
}
\arguments{
\item{data_directory}{is a relative path to specific folder where all CSV files must be loaded}
}
\description{
Gets RAW csv for IRR comparison from Azure
}
