% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_csv_from_azure.R
\name{get_raw_data_from_azure}
\alias{get_raw_data_from_azure}
\title{Get CSV Data from Azure}
\usage{
get_raw_data_from_azure(
  blob_directory = "source-data",
  filename = "pm_fund_info.csv",
  download_to_directory = "./",
  read_csv = TRUE
)
}
\arguments{
\item{blob_directory}{is the folder within the blob}

\item{filename}{is the filename within the blob's folder}

\item{download_to_directory}{is the relative path where the data will be downloaded and stored to}

\item{read_csv}{if the file is a CSV and you want it to be read, it will be returned as a tibble}
}
\description{
Gets RAW csv for IRR comparison from Azure
}
