% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ssbt_composite.R
\name{get_ssbt_composite}
\alias{get_ssbt_composite}
\title{Query ssbt_composite table from database}
\usage{
get_ssbt_composite(..., con = AZASRS_DATABASE_CONNECTION())
}
\arguments{
\item{...}{are for filtering the tables, can take values: category, portfolio, asset_class, sub_portfolio, previous_saa, sponsor, saa_benchmark, imp_benchmark, ticker, city}
}
\value{
Returns a tibble of all of the data
}
\description{
Query ssbt_composite table from database
}
\examples{
get_ssbt_composite(asset_class == 'Equities')
}
