% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ssbt_composite_bor_daily.R
\name{get_ssbt_composite_bor_daily}
\alias{get_ssbt_composite_bor_daily}
\title{Get SSBT composite book of record daily}
\usage{
get_ssbt_composite_bor_daily(con = AZASRS_DATABASE_CONNECTION(),
  return_tibble = TRUE)
}
\arguments{
\item{...}{filtering expressions. i.e. effective_date >= '2018-01-01'}
}
\value{
Returns a tibble with all composite data but can be filtered.
}
\description{
Gets composite book of record daily
}
\examples{
get_ssbt_composite_bor_daily(effective_date >= '2018-01-01')
# A tibble: 4,160 x 10
# ssbt_composite_in~ effective_date   beginning_market~ ending_market_v~ net_cash_flow daily_return ssbt_composite_~
#   <int>               <chr>              <dbl>            <dbl>         <dbl>        <dbl>           <chr>
#    24               2018-06-30         1467349533.      1482968778.          0       0.0106          Comp1
#    24               2018-07-02         1482974842.      1463381621.          0       0               Comp2
# ... with 4,150 more rows, and 3 more variables: ssbt_composite_description <chr>,
#  ssbt_composite_short_description <chr>, benchmark_info_id <int>
}
