% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ssbt_composite_bor_monthly.R
\name{get_ssbt_composite_bor_monthly}
\alias{get_ssbt_composite_bor_monthly}
\title{Get SSBT composite book of record monthly}
\usage{
get_ssbt_composite_bor_monthly(con = AZASRS_DATABASE_CONNECTION(),
  return_tibble = TRUE)
}
\arguments{
\item{...}{filtering expressions. i.e. effective_date >= '2018-01-01'}
}
\value{
Returns a tibble with all composite data but can be filtered.
}
\description{
Gets composite book of record monthly
}
\examples{
get_ssbt_composite_bor_monthly(effective_date >= '2018-01-01')
# A tibble: 4,160 x 10
# ssbt_composite_in~ effective_date   beginning_market~ ending_market_v~ net_cash_flow monthly_return fiscal_ytd_retu~
#   <int>               <chr>              <dbl>            <dbl>         <dbl>        <dbl>           <chr>
#    24               2018-12-31         1467349533.      1482968778.          0       0.0106          0
#    25               2018-12-31         1482974842.      1463381621.          0       0               0
# ... with 13 more rows, and 4 more variables: ssbt_composite_id <chr>, ssbt_composite_description <chr>,
#   ssbt_composite_short_description <chr>, benchmark_info_id <int>
}
