% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ssbt_composite_info.R
\name{get_ssbt_composite_info}
\alias{get_ssbt_composite_info}
\title{Query ssbt_composite_info table from database}
\usage{
get_ssbt_composite_info(
  con = AZASRS_DATABASE_CONNECTION(),
  return_tibble = TRUE
)
}
\arguments{
\item{...}{are for filtering the tables, can take values: category, portfolio, asset_class, sub_portfolio}
}
\value{
Returns a tibble of all of the data
}
\description{
Query ssbt_composite_info table from database
}
\examples{
get_ssbt_composite_info(asset_class == 'Equities', category == 'Large')
}
