% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_info_tbl.R
\name{get_tbl_info}
\alias{get_tbl_info}
\title{Query major info tables from database}
\usage{
get_tbl_info(tbl_name, ..., con = AZASRS_DATABASE_CONNECTION())
}
\arguments{
\item{tbl_name}{is the name of the major info table (choose one): pm_fund_info, ssbt_composite_info, account_info}

\item{...}{are for filtering the tables, can take values: category, portfolio, asset_class, sub_portfolio}
}
\value{
Returns a tibble of all of the data
}
\description{
Query major info tables from database
}
\examples{
get_tbl_info('pm_fund_info', sub_portfolio = 'Equity')
}
