% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_variables.R
\name{get_valdate}
\alias{get_valdate}
\title{Get valdate}
\usage{
get_valdate()
}
\arguments{
\item{db_con}{database connection}
}
\value{
valdate as character
}
\description{
Finds valdate from fundinfo table
}
