% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{LOAD_GLOBAL_DATA_PRIVATE_MARKETS}
\alias{LOAD_GLOBAL_DATA_PRIVATE_MARKETS}
\title{LOAD global data for private markets (all data from db)}
\usage{
LOAD_GLOBAL_DATA_PRIVATE_MARKETS(con = AZASRS_DATABASE_CONNECTION())
}
\arguments{
\item{con}{object from AZASRS_DATABASE_CONNECTION()}
}
\value{
list of ALL PM data
}
\description{
Create an object in the environment containing all underlying data for PM calcs
}
