% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_grouped_irrs.R
\name{merge_nav_and_cf}
\alias{merge_nav_and_cf}
\title{Merge NAV and CF}
\usage{
merge_nav_and_cf(.nav_data, .cf_data, end_date, cash_adjusted, pm_fund_info)
}
\arguments{
\item{.nav_data}{is from get_pm_nav_daily() through convert_start_date to negative()}

\item{.cf_data}{is from get_pm_cash_flow_daily() through filter_cf_between_dates}

\item{end_date}{is a string (format yyyy-dd-mm)}

\item{cash_adjusted}{is TRUE / FALSE to allow for cash adjusted NAV at next quarter}

\item{pm_fund_info}{from get_pm_fund_info()}
}
\description{
Merging both NAV and CF and handle complexity of rollup aggregation
}
