% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience functions.R
\name{modify_365_zoo}
\alias{modify_365_zoo}
\title{For the express use within get_variables functions}
\usage{
modify_365_zoo(data_frame, return_zoo, convert_365)
}
\description{
For the express use within get_variables functions
}
