% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_queries.R
\name{pull_benchmark}
\alias{pull_benchmark}
\title{Query benchmark data}
\usage{
pull_benchmark()
}
\arguments{
\item{db_con}{database connection}
}
\value{
dataframe of shortname, longname, date, price (adds in log price of ODCE as Fixed8)
}
\description{
SELECT statement to retrieve all historical_mv data
}
