% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_queries.R
\name{pull_cashflow}
\alias{pull_cashflow}
\title{Query cashflow data}
\usage{
pull_cashflow()
}
\arguments{
\item{db_con}{database connection}
}
\value{
cashflow dataframe (date, shortname, amount)
}
\description{
SELECT statement to retrieve all cashflow data
}
