% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_queries.R
\name{pull_categories}
\alias{pull_categories}
\title{Query category data}
\usage{
pull_categories()
}
\arguments{
\item{db_con}{database connection}
}
\value{
category dataframe **TODO: way too much info, need to narrow this down to important stuff
}
\description{
SELECT statement to retrieve all catetory data
}
