% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_queries.R
\name{pull_fundinfo}
\alias{pull_fundinfo}
\title{Query fundinfo data}
\usage{
pull_fundinfo()
}
\arguments{
\item{db_con}{database connection}
}
\value{
fundinfo dataframe columns requested back **TODO: currently retrieves all, need to narrow this down... should it be user feedback?
}
\description{
SELECT statement to retrieve all fundinfo data
}
