% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_queries.R
\name{pull_historicalmv}
\alias{pull_historicalmv}
\title{Query historical market value data}
\usage{
pull_historicalmv(db_con = db_connect())
}
\arguments{
\item{db_con}{database connection}
}
\value{
dataframe of shortname, date, amount
}
\description{
SELECT statement to retrieve all historical_mv data
}
