% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_queries.R
\name{pull_nav}
\alias{pull_nav}
\title{Query nav data}
\usage{
pull_nav()
}
\arguments{
\item{db_con}{database connection}

\item{from_date}{starting date ( >= ) at which data must be more recent than}
}
\value{
nav dataframe (date, shortname, amount)
}
\description{
SELECT statement to retrieve all nav data
}
