% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_queries.R
\name{raw_sql_query}
\alias{raw_sql_query}
\title{Query cashflow data}
\usage{
raw_sql_query(sql_query)
}
\arguments{
\item{sql_query}{SQL statement}
}
\value{
results from a SQL query as a data frame
}
\description{
SELECT statement of your choosing
}
