% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolling_p2p_irrs.R
\name{rolling_p2p_irr}
\alias{rolling_p2p_irr}
\title{BUILD PRIVATE MARKET P2P IRR TIBBLE}
\usage{
rolling_p2p_irr(end_date = get_value_date(),
  con = AZASRS_DATABASE_CONNECTION())
}
\arguments{
\item{end_date}{is the last date you would like the IRR to be calculated to
Should be in string format: 'yyyy-mm-dd'}

\item{start_date}{is the beginning date you would like the IRR to be calculated from
Should be in string format: 'yyyy-mm-dd'}
}
\description{
Creates a tibble of all Private Market IRRs found within the pm_fund_info table
}
