% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_raw_data_to_azure.R
\name{send_raw_data_to_azure}
\alias{send_raw_data_to_azure}
\title{Upload Data to Azure}
\usage{
send_raw_data_to_azure(data, blob_directory = "calculated-data", filename)
}
\arguments{
\item{data}{object to be converted to .rds and uploaded}

\item{blob_directory}{is the folder within the blob}

\item{filename}{is the filename within the blob's folder}
}
\description{
UPLOADS RAW data to Azure
}
