% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_globals.R
\docType{data}
\name{SHOW_ALL_TABLES}
\alias{SHOW_ALL_TABLES}
\title{List all tables and views in database}
\format{An object of class \code{character} of length 32.}
\usage{
SHOW_ALL_TABLES
}
\description{
Aids in displaying table names, simply add tbl_ in front of the name to access the function that accesses the table
}
\keyword{datasets}
