% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_grouped_pm_cash_flow.R
\name{test_for_no_aggregation}
\alias{test_for_no_aggregation}
\title{Tests to see if (...) is aggregating or not}
\usage{
test_for_no_aggregation(...)
}
\arguments{
\item{...}{is from filter_dates()}
}
\description{
Aggregation will change how funds are filtered, returns TRUE if not aggregated
}
