% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{test_is_not_rollup}
\alias{test_is_not_rollup}
\title{Test ... for rollup}
\usage{
test_is_not_rollup(...)
}
\arguments{
\item{...}{should be a column title from pm_fund_info (or multiple)}
}
\description{
TRUE is NOT a rollup, FALSE is a rollup (i.e. grouped beyond fund_id)
}
