% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trailing_period_calculations.R
\name{trailing_period_irr}
\alias{trailing_period_irr}
\title{Trailing Period IRR}
\usage{
trailing_period_irr(shortname = "Total PE", portfolio_name = "PE",
  comparison_years = c(1, 3, 5), benchmark_type = "benchmark",
  benchmark_symbol = "DEFAULT_FROM_FUNDINFO")
}
\arguments{
\item{shortname}{name of total portfolio from nav and cashflow tables}

\item{portfolio_name}{name of portfolio from fundinfo table}

\item{comparison_years}{trailing years, default is c(1,3,5)}

\item{benchmark_type}{from fundinfo table: benchmark, stype, sector}

\item{benchmark_symbol}{benchmark to compare against - this overrides benchmark_type}

\item{db_con}{database connection}
}
\value{
dataframe of years/dates, IRR benchmark, IRR priv
}
\description{
Calculates trailing period IRR
}
