% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_globals.R
\name{UPDATE_DATABASE}
\alias{UPDATE_DATABASE}
\title{Updates database with AZURE files}
\usage{
UPDATE_DATABASE(filename, development = 0, local_azure_functions = FALSE)
}
\description{
To be used with individual files. Note: the first connection takes a long time to spin up the server, then should work quickly if more files are necessary.
}
\examples{
UPDATE_DATABASE("pm_fund_info.csv")
}
