% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_globals.R
\docType{data}
\name{AZASRS_TEST_DATA_DIRECTORY}
\alias{AZASRS_TEST_DATA_DIRECTORY}
\title{AZASRS TEST DIRECTORY}
\format{An object of class \code{character} of length 1.}
\usage{
AZASRS_TEST_DATA_DIRECTORY
}
\description{
Data for testing purposes
}
\keyword{datasets}
