% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_grouped_irrs.R
\name{build_grouped_irrs}
\alias{build_grouped_irrs}
\title{BUILDS IRR's}
\usage{
build_grouped_irrs(
  ...,
  con = AZASRS_DATABASE_CONNECTION(),
  nav_daily = get_pm_nav_daily(con = con, return_tibble = FALSE),
  cash_flow_daily = get_pm_cash_flow_daily(con = con, return_tibble = FALSE),
  start_date = "2017-12-31",
  end_date = get_value_date()
)
}
\arguments{
\item{...}{is the grouping set of variable(s) requested}

\item{start_date}{is the beginning date you would like the IRR to be calculated from
Should be in string format: 'yyyy-mm-dd'}

\item{end_date}{is the last date you would like the IRR to be calculated to
Should be in string format: 'yyyy-mm-dd'}
}
\description{
Builds cash flow list properly and adds NAV where required
}
