% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_grouped_irrs.R
\name{build_grouped_irrs}
\alias{build_grouped_irrs}
\title{Create rollup of IRR based off of dates}
\usage{
build_grouped_irrs(
  start_date,
  end_date,
  itd,
  cash_adjusted,
  nav_daily,
  cf_daily,
  pm_fund_info,
  ...
)
}
\arguments{
\item{start_date}{is the start date of analysis}

\item{end_date}{is the cutoff date of analysis}

\item{itd}{is a boolean that determines whether itd (inception to date) is included}

\item{cash_adjusted}{is a boolean that determines whether cash is adjusted}

\item{nav_daily}{is the object of get_pm_nav_daily()}

\item{cf_daily}{is the object of get_pm_cash_flow_daily()}

\item{pm_fund_info}{is the object of get_pm_fund_info()}

\item{...}{aggregation choices from from pm_fund_info (i.e. pm_fund_portfolio, pm_fund_category, pm_fund_id)}

\item{.data}{is from clean_nav_cf()}
}
\value{

}
\description{
Build IRR for any rollup and date range, including ITD and cash adjusted
}
\examples{
# Example use case
nav <- get_pm_nav_daily() \%>\% dplyr::filter(nav != 0)
cf <- get_pm_cash_flow_daily() \%>\% dplyr::filter(cash_flow != 0)
pm_fund_info <- get_pm_fund_info()
start_date <- "2019-09-30"
end_date <- "2019-12-31"
itd <- FALSE
cash_adjusted <- FALSE
final_data <- build_grouped_irrs(
  start_date = start_date, end_date = end_date, itd = itd,
  cash_adjusted = cash_adjusted, pm_fund_info = pm_fund_info,
  pm_fund_portfolio, pm_fund_category_description
)
}
