% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_grouped_pm_cash_flow.R
\name{build_grouped_pm_cash_flow}
\alias{build_grouped_pm_cash_flow}
\title{Build PM "Cash Flow" and utilizing NAV to simulate where necessary}
\usage{
build_grouped_pm_cash_flow(
  start_date,
  end_date,
  itd,
  cash_adjusted,
  nav_daily,
  cf_daily,
  bench_daily,
  bench_relationships,
  pm_fund_info,
  ...
)
}
\arguments{
\item{...}{aggregation choices from from pm_fund_info (i.e. pm_fund_portfolio, pm_fund_category, pm_fund_id)}

\item{.data}{is from clean_nav_cf()}
}
\description{
Combine cash flow and NAV data to create a "cash flow" for private market calculations
}
\examples{
# Example use case
nav = get_pm_nav_daily() \%>\% dplyr::filter(nav != 0)
cf = get_pm_cash_flow_daily() \%>\% dplyr::filter(cash_flow != 0)
pm_fund_info = get_pm_fund_info()
start_date = '2019-09-30'
end_date = '2019-12-31'
itd = FALSE
cash_adjusted = FALSE
final_data = build_grouped_pm_cash_flow(start_date = start_date, end_date = end_date, itd = itd,
                                        cash_adjusted = cash_adjusted, pm_fund_info = pm_fund_info,
                                        pm_fund_portfolio, pm_fund_category_description)
}
