% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_grouped_pm_cash_flow.R
\name{build_grouped_pm_cash_flow}
\alias{build_grouped_pm_cash_flow}
\title{Build PM "Cash Flow" and utilizing NAV to simulate where necessary}
\usage{
build_grouped_pm_cash_flow(
  ...,
  con = AZASRS_DATABASE_CONNECTION(),
  start_date = "2019-06-30",
  end_date = get_value_date(con = con),
  itd = FALSE,
  cash_adjusted = FALSE,
  nav_daily = get_pm_nav_daily(con = con),
  cf_daily = get_pm_cash_flow_daily(con = con),
  bench_daily = get_benchmark_daily_index(con = con, benchmark_type = "PVT",
    return_tibble = TRUE),
  bench_relationships = get_benchmark_fund_relationship(con = con, bench_type = "PVT",
    return_tibble = TRUE),
  pm_fund_info = get_pm_fund_info(con = con)
)
}
\arguments{
\item{...}{aggregation choices from from pm_fund_info (i.e. pm_fund_portfolio, pm_fund_category, pm_fund_id)}

\item{start_date}{is the start date of analysis}

\item{end_date}{is the cutoff date of analysis, typically a value date}

\item{itd}{is a boolean that determines whether itd (incpetion to date) is included, overrides start date}

\item{cash_adjusted}{is a boolean that determines if cash is adjusted NAV should be used}

\item{nav_daily}{is the object of get_pm_nav_daily()}

\item{cf_daily}{is the object of get_pm_cash_flow_daily()}

\item{bench_daily}{is the object of get_benchmark_daily_index()}

\item{bench_relationships}{is the object of get_benchmark_fund_relationship()}

\item{pm_fund_info}{is the object of get_pm_fund_info()}

\item{.data}{is from clean_nav_cf()}
}
\description{
Combine cash flow and NAV data to create a "cash flow" for private market calculations
}
