% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_grouped_pm_metrics.R
\name{build_grouped_pm_metrics}
\alias{build_grouped_pm_metrics}
\title{Create rollup of IRR based off of dates}
\usage{
build_grouped_pm_metrics(
  ...,
  con = AZASRS_DATABASE_CONNECTION(),
  start_date = "2016-12-31",
  end_date = get_value_date(con = con),
  itd = FALSE,
  cash_adjusted = FALSE,
  nav_daily = get_pm_nav_daily(con = con),
  cf_daily = get_pm_cash_flow_daily(con = con),
  bench_daily_index = get_benchmark_daily_index(con = con, benchmark_type = "PVT",
    return_tibble = TRUE),
  bench_relationships = get_benchmark_fund_relationship(con = con, bench_type = "PVT",
    return_tibble = TRUE),
  pm_fund_info = get_pm_fund_info(con = con)
)
}
\arguments{
\item{...}{aggregation choices from from pm_fund_info (i.e. pm_fund_portfolio, pm_fund_category, pm_fund_id)}

\item{start_date}{is the start date of analysis}

\item{end_date}{is the cutoff date of analysis}

\item{itd}{is a boolean that determines whether itd (incpetion to date) is included, overrides start_date}

\item{cash_adjusted}{is a boolean that determines if NAV is cash adjusted}

\item{nav_daily}{is the object of get_pm_nav_daily()}

\item{cf_daily}{is the object of get_pm_cash_flow_daily()}

\item{bench_daily_index}{is the object of get_benchmark_daily_index()}

\item{bench_relationships}{is the object of get_benchmark_fund_relationship()}

\item{pm_fund_info}{is the object of get_pm_fund_info()}
}
\description{
Build IRR for any rollup and date range, including ITD and cash adjusted
}
