% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_privm_metrics.R
\name{build_privm_metrics}
\alias{build_privm_metrics}
\title{Build a tibble of major private market metrics including: IRR, DPI, TVPI, PME -- Assumes benchmark type is PVT}
\usage{
build_privm_metrics(
  ...,
  con = AZASRS_DATABASE_CONNECTION(),
  nav_daily = get_pm_nav_daily(con = con, return_tibble = FALSE),
  cf_daily = get_pm_cash_flow_daily(con = con, return_tibble = FALSE),
  benchmark_daily = build_benchmark_fv_index_factor(con = con, return_tibble = FALSE),
  pmfi = get_pm_fund_info(con = con, return_tibble = FALSE),
  start_date = "1900-01-01",
  value_date = get_value_date(con = con),
  pcap_date = as.character(lubridate::today())
)
}
\arguments{
\item{...}{grouping variables (pm_fund_id, pm_fund_portfolio, etc.)}

\item{con}{is a database connection object from AZASRS::AZASRS_DATABASE_CONNECTION()}

\item{nav_daily}{is the object of get_pm_nav_daily()}

\item{cf_daily}{is the object of get_pm_cash_flow_daily()}

\item{benchmark_daily}{is the object of build_benchmark_fv_index_factor()}

\item{pmfi}{is the object of get_pm_fund_info()}

\item{start_date}{is the start date of analysis}

\item{value_date}{is the object of get_value_date()}

\item{pcap_date}{is the cutoff date of analysis}
}
\value{
Returns a tibble with grouping variables and all of their respective metrics
}
\description{
Build a tibble of major private market metrics including: IRR, DPI, TVPI, PME -- Assumes benchmark type is PVT
}
\examples{
build_privm_metrics(pm_fund_portfolio, pm_fund_id)
# pm_fund_portfolio pm_fund_id    irr     dpi   tvpi appreciation     pme irr_pcap
# <chr>             <chr>       <dbl>   <dbl>  <dbl>        <dbl>   <dbl>    <dbl>
# 1 Credit            Fund1     0.0919   0.372   1.15   437251492.   1.05        0.11
# 2 Credit            Fund2     0.159    0       1.11   562211000    1.10        NA
# 3 Credit            Fund3     0.0247   0       1.10   262390142    0.952       NA
# 4 Credit            Fund4     0.0673   0.135   1.17   173259780    1.06        NA
}
