% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_finance_basics.R
\name{calc_dpi}
\alias{calc_dpi}
\title{DPI}
\usage{
calc_dpi(distributions, contributions)
}
\arguments{
\item{distributions}{are positive cash flow}

\item{contributions}{are negative cash flow}
}
\description{
Calculates DPI that is NOT based off of asrsMethods library. asrsMethods needs to be updated because it does not reflect TVPI as Kerry calculates.
}
