% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_finance_basics.R
\name{calc_dva}
\alias{calc_dva}
\title{DVA}
\usage{
calc_dva(cash_flow, fv_index_factors)
}
\arguments{
\item{cash_flow}{is the cash flow tibble}

\item{fv_index_factors}{are a c() of future value data points based off of index value divided by first index value (i.e. 1.18, 1.11, 1.05, 1.00)}
}
\description{
Calculates DVA that is NOT based off of asrsMethods library.
}
