% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_finance_basics.R
\name{calc_irr_from_df}
\alias{calc_irr_from_df}
\title{IRR from dataframe}
\usage{
calc_irr_from_df(df)
}
\arguments{
\item{df}{is a dataframe with columns named cash_flow and dates}
}
\description{
Abstraction of calc_irr to use with dataframe
}
