% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_account_info.R
\name{get_account_info}
\alias{get_account_info}
\title{Query account info table from database}
\usage{
get_account_info(..., con = AZASRS_DATABASE_CONNECTION())
}
\arguments{
\item{...}{are for filtering the tables, can take values: category, portfolio, asset_class, sub_portfolio, previous_saa, sponsor, saa_benchmark, imp_benchmark, ticker, city}
}
\value{
Returns a tibble of all of the data
}
\description{
Query account info table from database
}
\examples{
get_account_info(asset_class == 'Equities', category == 'Large')
}
