% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_benchmark_daily_index.R
\name{get_benchmark_daily_index}
\alias{get_benchmark_daily_index}
\title{Get all benchmark_daily_index}
\usage{
get_benchmark_daily_index(
  con = AZASRS_DATABASE_CONNECTION(),
  benchmark_type = "PVT",
  all_bench_types = FALSE,
  return_tibble = FALSE
)
}
\arguments{
\item{con}{is a database connection object from AZASRS::AZASRS_DATABASE_CONNECTION()}

\item{all_bench_types}{is a boolean that determines whether or not all bench types are returned}

\item{return_tibble}{is a boolean that determines whether or not a tibble is returned instead}

\item{bench_type}{is the type of benchmark where default is 'PVT' which is private}
}
\value{
Returns a tibble or database object.
}
\description{
Gets benchmark index data by day and only allows one bench_type (i.e. SAA). Joins data from pm_fund_info, benchmark_info and filters by benchmark type. Caution: returns A LOT of data, filter for better performance.
}
\examples{
get_benchmark_daily_index(return_tibble = TRUE)
# A tibble: 76,693 x 4
# benchmark_info_id effective_date index_value benchmark_id
# <int>                <date>               <dbl>           <chr>
#  53                 2004-01-02            1             LSTA+250
#  53                 2004-01-03            1.00          LSTA+250
#  53                 2004-01-04            1.00          LSTA+250
# ... with 76,690 more rows
}
