% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_variables.R
\name{get_filtered_benchmark}
\alias{get_filtered_benchmark}
\title{Query filtered benchmark data}
\usage{
get_filtered_benchmark(shortname)
}
\arguments{
\item{db_con}{database connection}
}
\value{
dataframe of shortname, longname, date, price (adds in log price of ODCE as Fixed8)
}
\description{
SELECT statement to retrieve benchmark data filtered by shortname
}
