% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_info_tbl.R
\name{get_info_tbl}
\alias{get_info_tbl}
\title{Query major info tables from database}
\usage{
get_info_tbl(tbl_name, ..., con = AZASRS_DATABASE_CONNECTION())
}
\arguments{
\item{tbl_name}{is the name of the major info table (choose one): pm_fund_info, ssbt_composite_info, account_info}

\item{...}{are for filtering the tables, can take values: category, portfolio, asset_class, sub_portfolio}
}
\value{
A table object from dplyr
}
\description{
Query major info tables from database
}
\examples{
get_info_tbl('pm_fund_info', sub_portfolio='Equity')
}
