% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pm_fund_info.R
\name{get_pm_fund_info}
\alias{get_pm_fund_info}
\title{Get all pm_fund_info}
\usage{
get_pm_fund_info(con = AZASRS_DATABASE_CONNECTION(), return_tibble = TRUE)
}
\arguments{
\item{...}{filtering expressions. i.e. pm_fund_portfolio == 'Credit'}
}
\value{
Returns a tibble with all pm_fund_info metadata.
}
\description{
A view to get all private market fund info, can be filtered. By default, SQL --> SELECT * FROM all_pm_fund_info;
}
\examples{
get_pm_fund_info(pm_fund_portfolio == 'Credit')
# A tibble: 37 x 28
# pm_fund_info_id pm_fund_id pm_fund_description  pm_fund_portfol~ pm_fund_sponsor~ pm_fund_city_id pm_fund_sector_~
# <int> <chr>      <chr>                           <int>            <int>           <int>            <int>
# 1               1 Fund1    FundDesc1                1                1               1                1
# 2               2 Fund2    FundDesc2                1                1               1                1
# 3               3 Fund3    FundDesc3                1                1               1                1
# ... with 27 more rows, and 21 more variables: vintage <int>, commit <int>, unfunded <dbl>, legacy <chr>,
#   specialist <chr>, invest_end <chr>, term_end <chr>, extension <dbl>, ext_time <dbl>, ext_used <dbl>,
#   fee_cat <chr>, consultant <chr>, adv_board <int>, obsvr <int>, fund_size_m <dbl>, pm_fund_category <chr>,
#   pm_fund_category_description <chr>, pm_fund_city <chr>, pm_fund_portfolio <chr>, pm_fund_sector <chr>,
#   pm_fund_sponsor <chr>
}
