% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pm_fund_info.R
\name{get_pm_fund_info}
\alias{get_pm_fund_info}
\title{Get all pm_fund_info}
\usage{
get_pm_fund_info(
  con = AZASRS_DATABASE_CONNECTION(),
  add_benchmark = FALSE,
  return_tibble = TRUE
)
}
\arguments{
\item{con}{is a database connection object from AZASRS::AZASRS_DATABASE_CONNECTION()}

\item{return_tibble}{is a boolean that determines whether or not a tibble is returned instead}

\item{add_benckmark}{is a boolean that appends multiple benchmarks as their own columns. i.e. PVT_Benchmark, SAA_Benchmark, etc.}
}
\value{
Returns a tibble or SQL result with all pm_fund_info metadata.
}
\description{
A view to get all private market fund info, can be filtered. By default, SQL --> SELECT * FROM all_pm_fund_info;
}
\examples{
get_pm_fund_info()
# A tibble: 282 x 26
# pm_fund_info_id pm_fund_id pm_fund_descrip… pm_fund_common_… vintage commit unfunded legacy specialist
# <int> <chr>      <chr>                   <chr>              <int>   <int>    <int>    <chr>  <chr>
# 1      Hgh19     AP Mezzanine Pa…      HPS Mezz 2019         2019    6.00e8  3.95e8     A    " "
# 2      HghBr     AP Mezzanine Pa…      HPS Mezz 2            2013    2.00e8  1.30e7     A    " "
# 3      HghBr3    AP Mezzanine Pa…      HPS Mezz 3            2016    5.00e8  9.85e7     A    " "
}
