% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ssbt_composite_info.R
\name{get_ssbt_composite_info}
\alias{get_ssbt_composite_info}
\title{Query ssbt_composite_info table from database}
\usage{
get_ssbt_composite_info(
  con = AZASRS_DATABASE_CONNECTION(),
  return_tibble = TRUE
)
}
\arguments{
\item{con}{is a database connection object from AZASRS::AZASRS_DATABASE_CONNECTION()}

\item{return_tibble}{is a boolean that determines whether or not a tibble is returned instead}
}
\value{
Returns a tibble or database object
}
\description{
Gets all ssbt composite information.
}
\examples{
get_ssbt_composite_info()
# A tibble: 36 x 5
# ssbt_composite_info… ssbt_composite_id ssbt_composite_description  ssbt_composite_short_des… benchmark_info_…
# <int>                <chr>               <chr>                <chr>            <int>
#  1                 ASRSA001           TOTAL PLAN            Total Fund          NA
#  2                 ASRSA007           TOTAL EQUITY          Equity              NA
#  3                 ASRSA007           TOTAL PUBLIC EQUITY   Public Equity       NA
# ... with 33 more rows
}
