% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_total_fund_value.R
\name{get_total_fund_value}
\alias{get_total_fund_value}
\title{Get total fund value}
\usage{
get_total_fund_value(
  value_date = get_value_date(),
  con = AZASRS_DATABASE_CONNECTION()
)
}
\arguments{
\item{value_date}{grabs the date from the get_value_data() function}

\item{con}{is a database connection object from AZASRS::AZASRS_DATABASE_CONNECTION()}
}
\value{
Value from ssbt_composite_book_of_record_daily by date selected (default = valdate).
}
\description{
Finds the most recent fund value for TOTAL FUND. Can be filtered to select by individual date.
}
\examples{
get_total_fund_value(value_date = "2018-06-30")
# 39656200393
}
